feat: move composables to commons (batch 8)
Move 15 UI files from amethyst/ui/ to commons module with backward-compatible
typealiases and wrapper functions at original locations.

Files moved (zero non-commons deps):
- ToastMsg, ResourceToastMsg, StringToastMsg, ThrowableToastMsg (toast system)
- BookmarkType (enum)
- IExpiration, IZapRaiser (interfaces)
- TranslationConfig (data class)
- MediaUploadTracker (upload state)
- ServerName, ServerType, DEFAULT_MEDIA_SERVERS (media servers)
- DraftTagState (draft tag management)
- AdditiveComplexFeedFilter (feed filter base)
- ShorterTopAppBar (top bar composable)
- DisappearingFloatingButton (FAB composable)

Also fixes smart-cast issues in TranslatableRichTextViewer and
DisplayErrorMessages for cross-module property access.
