+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/components/TranslationConfig.kt
@@ -20,12 +20,5 @@
  */
 package com.vitorpamplona.amethyst.ui.components
 
-import androidx.compose.runtime.Immutable
-
-@Immutable
-data class TranslationConfig(
-    val result: String?,
-    val sourceLang: String?,
-    val targetLang: String?,
-    val showOriginal: Boolean,
-)
+// Backward compatibility - moved to commons
+typealias TranslationConfig = com.vitorpamplona.amethyst.commons.ui.components.TranslationConfig
