+++ commons/src/commonMain/kotlin/com/vitorpamplona/amethyst/commons/ui/note/creators/draftTags/DraftTagState.kt
@@ -18,7 +18,43 @@
  * AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.vitorpamplona.amethyst.ui.note.creators.zapraiser
+package com.vitorpamplona.amethyst.commons.ui.note.creators.draftTags
 
-// Backward compatibility - moved to commons
-typealias IZapRaiser = com.vitorpamplona.amethyst.commons.ui.note.creators.zapraiser.IZapRaiser
+import androidx.compose.runtime.Stable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.setValue
+import kotlinx.coroutines.FlowPreview
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.debounce
+import kotlinx.coroutines.flow.update
+import kotlin.uuid.ExperimentalUuidApi
+import kotlin.uuid.Uuid
+
+@Stable
+class DraftTagState {
+    var current: String by mutableStateOf(newTag())
+    var usedDraftTags by mutableStateOf(setOf(current))
+
+    private val _versions = MutableStateFlow(0)
+
+    @OptIn(FlowPreview::class)
+    val versions = _versions.debounce(1000)
+
+    @OptIn(ExperimentalUuidApi::class)
+    fun newTag() = Uuid.random().toString()
+
+    fun rotate() {
+        set(newTag())
+        _versions.update { 0 }
+    }
+
+    fun set(existingTag: String) {
+        current = existingTag
+        usedDraftTags += existingTag
+    }
+
+    fun newVersion() {
+        _versions.update { it + 1 }
+    }
+}
